requirements.txt
torch_scatter==2.0.9
tqdm==4.62.3
torch==1.10.0
numpy==1.20.3
torch_sparse==0.6.12
torch_geometric==2.0.2
PyYAML==6.0
scikit_learn==1.0.2
